[ADMINAPI-1302] - Ed-Orgs-Endpoints - Creates service interface and service implementation (#246)
* Ed-Orgs-Endpoints - Creates service interface and service implementation

* Update Application/EdFi.Ods.AdminApi/Features/Vendors/ReadVendor.cs


* Update Application/EdFi.Ods.AdminApi/Infrastructure/Services/EducationOrganizationService/EducationOrganizationService.cs


* Update Application/EdFi.Ods.AdminApi/Infrastructure/Services/EducationOrganizationService/EducationOrganizationService.cs


* Update Application/EdFi.Ods.AdminApi/Infrastructure/Services/EducationOrganizationService/EducationOrganizationService.cs


* Update Application/EdFi.Ods.AdminApi/Infrastructure/Database/IAdminApiUserContext.cs


* Update Application/EdFi.Ods.AdminApi/Infrastructure/WebApplicationBuilderExtensions.cs


* Refactor tenant processing in EducationOrganizationService

Simplifies tenant configuration handling by filtering and processing tenants with valid configurations in a separate method. Updates database engine comparisons to use DatabaseEngineEnum and passes CancellationToken.None to SaveChangesAsync for consistency.

* Refactor EducationOrganizationService to use AdminApiDbContext

Updated EducationOrganizationService to use AdminApiDbContext for EducationOrganization operations instead of IAdminApiUserContext. Registered new context types in dependency injection and added EducationOrganizations DbSet to AdminApiDbContext. This change improves separation of concerns and aligns context usage with entity responsibilities.

* Remove AdminConsole related db context classes. Refactor the code to use IUserContext directly

* deletes unnecessary file OdsInstances.http.

---------

+++ Application/EdFi.Ods.AdminApi.Common/Infrastructure/Models/EducationOrganization.cs
@@ -10,11 +10,11 @@ public class EducationOrganization
     public int Id { get; set; }
     public int InstanceId { get; set; }
     public string InstanceName { get; set; } = string.Empty;
-    public int EducationOrganizationId { get; set; }
+    public long EducationOrganizationId { get; set; }
     public string NameOfInstitution { get; set; } = string.Empty;
     public string? ShortNameOfInstitution { get; set; }
     public string Discriminator { get; set; } = string.Empty;
-    public int? ParentId { get; set; }
+    public long? ParentId { get; set; }
     public string? OdsDatabaseName { get; set; }
     public DateTime LastRefreshed { get; set; }
     public DateTime? LastModifiedDate { get; set; }
